Move IBC files around
