+++ Makefile
@@ -18,4 +18,13 @@ init:
 tendermint-start:
 	tendermint start --home ~/.gaia-rs
 
-.PHONY: run run-debug test install init tendermint-start
+init-second:
+	./gaia-rs/scripts/init_second.sh
+
+tendermint-start-second:
+	tendermint start --home ~/.gaia-rs-second
+
+run-second:
+	cargo run -- run --home ~/.gaia-rs-second --address "127.0.0.1:26661" --rest-listen-addr "127.0.0.1:1318"
+
+.PHONY: run run-debug test install init tendermint-start init-second tendermint-start-second run-second
