+++ database/src/rocks.rs
@@ -25,7 +25,7 @@ impl Database for RocksDB {
     fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
         self.db
             .get(key)
-            .unwrap_or_else(|e| panic!("unrecoverable database error {}", e))
+            .unwrap_or_else(|e| panic!("unrecoverable database error {}", e)) //TODO: this is probably not the right thing to do when handling a abci or REST query. Perhaps we should have a must_get method?
     }
 
     fn put(&self, key: Vec<u8>, value: Vec<u8>) {
