+++ x/ibc/src/client/cli/OLD_tx/recover_client.rs
@@ -3,7 +3,7 @@ use ibc::{
     core::{client::types::proto::v1::MsgRecoverClient, host::types::identifiers::ClientId},
     primitives::Signer,
 };
-//use proto_messages::cosmos::ibc::{tx::MsgRecoverClient, types::core::host::identifiers::ClientId};
+
 
 #[derive(Args, Debug, Clone)]
 pub struct CliRecoverClient {
