+++ x/ibc/src/client/cli/mod.rs
@@ -1,3 +1,2 @@
 pub mod query;
-pub mod query_handler;
 pub mod tx;
