+++ x/ibc/src/ics02_client/client/cli/query/client_states.rs
@@ -1,6 +1,5 @@
 use clap::Args;
 use ibc::core::client::types::proto::v1::QueryClientStatesRequest;
-//use proto_messages::cosmos::ibc::types::core::client::context::types::proto::v1::QueryClientStatesRequest;
 
 pub(crate) const STATES_URL: &str = "/ibc.core.client.v1.Query/ClientStates";
