+++ x/ibc/src/types/genesis.rs
@@ -8,10 +8,3 @@ pub struct GenesisState {
     pub connection_genesis: ics03_connection::GenesisState,
     pub channel_genesis: ics04_channel::GenesisState,
 }
-
-// impl Default for GenesisState {
-//     fn default() -> Self {
-
-//         Self {}
-//     }
-// }
